fix: rename StackContent illustration + tip components simplify illustrations (please squash me)
