+++ apps/desktop/src/components/v3/StackDetails.svelte
@@ -1,20 +1,11 @@
-<script lang="ts" module>
-	export const previewModes = {
-		EmptyBranch: 'EmptyBranch',
-		SelectToPreview: 'SelectToPreview',
-		ViewingTips: 'ViewingTips',
-		ViewingPreview: 'ViewingPreview'
-	} as const;
-
-	export type PreviewMode = keyof typeof previewModes;
-</script>
-
 <script lang="ts">
 	import { isArchivedBranch, isStackedBranch } from './lib';
 	import ReduxResult from '$components/ReduxResult.svelte';
 	import Resizer from '$components/Resizer.svelte';
 	import Branch from '$components/v3/Branch.svelte';
-	import StackContentPlaceholder from '$components/v3/StackContentPlaceholder.svelte';
+	import StackContentIllustration, {
+		PreviewMode
+	} from '$components/v3/StackContentIllustration.svelte';
 	import { SETTINGS, type Settings } from '$lib/settings/userSettings';
 	import { StackService } from '$lib/stacks/stackService.svelte';
 	import { getContext, getContextStoreBySymbol } from '@gitbutler/shared/context';
@@ -37,14 +28,14 @@
 	const stackData = $derived(result.current.data?.[0]);
 
 	const stackContentMode = $derived.by<PreviewMode>(() => {
-		if (!stackData) return previewModes.EmptyBranch;
-		if (isArchivedBranch(stackData.state)) return previewModes.SelectToPreview;
+		if (!stackData) return PreviewMode.EmptyBranch;
+		if (isArchivedBranch(stackData.state)) return PreviewMode.SelectToPreview;
 
 		if (isStackedBranch(stackData.state) && stackData.state.subject.localAndRemote.length === 0) {
-			return previewModes.EmptyBranch;
+			return PreviewMode.EmptyBranch;
 		}
 
-		return previewModes.SelectToPreview;
+		return PreviewMode.SelectToPreview;
 	});
 </script>
 
@@ -82,7 +73,7 @@
 	</div>
 
 	<div class="stack__branch-content">
-		<StackContentPlaceholder mode={stackContentMode} />
+		<StackContentIllustration mode={stackContentMode} />
 	</div>
 </div>
